requirements.txt
# AutoPPM - Automated Portfolio Management Platform
# Core Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Zerodha Kite Connect Integration
kiteconnect==5.0.0

# Database
sqlalchemy==2.0.23

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
numpy==1.25.2

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging & Monitoring
loguru==0.7.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Time & Date
python-dateutil==2.8.2
pytz==2023.3

# JSON & Serialization
orjson==3.9.10

# Async Support
asyncio-mqtt==0.16.1
